fix: rename stackingBranch* components to stackingSeries*
