+++ apps/desktop/src/lib/branch/StackingSeriesHeaderContextMenu.svelte
